Make stop os specific
